+++ src/drivers/core.go
@@ -8,9 +8,41 @@ import (
 
 // Core provides the public API for the drivers subsystem.
 
-// GetActiveDriver returns the code hosting driver to use based on the git config.
+// CodeHostingDriver defines the structure of drivers
+// for the different code hosting services.
+type CodeHostingDriver interface {
+
+	// CanMergePullRequest returns whether or not MergePullRequest should be
+	// called when shipping. If true, also returns the default commit message
+	CanMergePullRequest(branch, parentBranch string) (canMerge bool, defaultCommitMessage string, pullRequestNumber int64, err error)
+
+	// GetNewPullRequestURL returns the URL of the page
+	// to create a new pull request online
+	GetNewPullRequestURL(branch, parentBranch string) string
+
+	// MergePullRequest merges the pull request through the hosting service api
+	MergePullRequest(MergePullRequestOptions) (mergeSha string, err error)
+
+	// GetRepositoryURL returns the URL where the given repository
+	// can be found online
+	GetRepositoryURL() string
+
+	// HostingServiceName returns the name of the code hosting service
+	HostingServiceName() string
+}
+
+// MergePullRequestOptions defines the options to the MergePullRequest function.
+type MergePullRequestOptions struct {
+	Branch            string
+	PullRequestNumber int64
+	CommitMessage     string
+	LogRequests       bool
+	ParentBranch      string
+}
+
+// Get returns the code hosting driver to use based on the git config.
 // nolint:interfacer  // for Gitea support later
-func GetActiveDriver(config *git.Configuration) (CodeHostingDriver, error) {
+func Get(config *git.Configuration) (CodeHostingDriver, error) {
 	driver := TryUseGithub(config)
 	if driver != nil {
 		return driver, nil
