--- src/drivers/merge_pull_request_options.go
@@ -1,10 +1,0 @@
-package drivers
-
-// MergePullRequestOptions defines the options to the MergePullRequest function.
-type MergePullRequestOptions struct {
-	Branch            string
-	PullRequestNumber int64
-	CommitMessage     string
-	LogRequests       bool
-	ParentBranch      string
-}
